refactor: 이벤트 핸들러 분리
